isolated and refactored dqn code
