--- README.md
@@ -1,25 +1,0 @@
-# Deep Reinforcement Learning In Action
-
-Code Snippets from the [Deep Reinforcement Learning in Action](https://www.manning.com/books/deep-reinforcement-learning-in-action) book from Manning, Inc
-
-## How this is Organized
-
-The code snippets, listings, and projects are all embedded in Jupyter Notebooks
-organized by chapter. Visit [http://jupyter.org/install](http://jupyter.org/install) for
-instructions on installing Jupyter Notebooks.
-
-We keep the original Jupyter Notebooks in their respective chapter folders. As we discover errata, we update notebooks in the Errata folder, so those notebooks are the most up-to-date in terms of errors corrected, but we keep the original Jupyter Notebooks to match the book code snippets.
-
-## Requirements
-
-In order to run many of the projects, you'll need at least the [NumPy](http://www.numpy.org/) library
-and [PyTorch](http://pytorch.org/).
-
-```
-pip install -r requirements.txt
-```
-
-## Contribute
-
-If you experience any issues running the examples, please file an issue.
-If you see typos or other errors in the book, please edit the [Errata.md](https://github.com/DeepReinforcementLearning/DeepReinforcementLearningInAction/blob/master/Errata.md) file and create a pull request.
