+++ agent/properties.py
@@ -1,0 +1,56 @@
+import numpy as np
+import torch
+from environment.Gridworld import Gridworld
+from IPython.display import clear_output
+import random
+from matplotlib import pylab as plt
+from collections import deque
+import copy
+
+l1 = 64
+l2 = 150
+l3 = 100
+l4 = 4
+
+mem_size = 1000
+batch_size = 200
+replay = deque(maxlen=mem_size)
+
+epochs = 5000
+losses = [] #A
+
+max_moves = 50
+h = 0
+sync_freq = 500 #A
+j=0
+
+model = torch.nn.Sequential(
+    torch.nn.Linear(l1, l2),
+    torch.nn.ReLU(),
+    torch.nn.Linear(l2, l3),
+    torch.nn.ReLU(),
+    torch.nn.Linear(l3,l4)
+)
+
+model2 = copy.deepcopy(model) #A
+model2.load_state_dict(model.state_dict()) #B|
+
+loss_fn = torch.nn.MSELoss()
+learning_rate = 1e-3
+optimizer = torch.optim.Adam(model.parameters(), lr=learning_rate)
+
+gamma = 0.9
+epsilon = 1.0
+learning_rate = 1e-3
+optimizer = torch.optim.Adam(model.parameters(), lr=learning_rate)
+
+gamma = 0.9
+epsilon = 1.0
+
+action_set = {
+    0: 'u',
+    1: 'd',
+    2: 'l',
+    3: 'r',
+}
+
