+++ common/test.py
@@ -1,0 +1,47 @@
+import numpy as np
+import torch
+from environment.Gridworld import Gridworld
+from IPython.display import clear_output
+import random
+from matplotlib import pylab as plt
+from agent.properties import *
+
+def test_model(model, mode='static', display=True):
+    i = 0
+    test_game = Gridworld(mode=mode)
+    state_ = test_game.board.render_np().reshape(1,64) + np.random.rand(1,64)/10.0
+    state = torch.from_numpy(state_).float()
+    if display:
+        print("Initial State:")
+        print(test_game.display())
+    status = 1
+    while(status == 1): #A
+        qval = model(state)
+        qval_ = qval.data.numpy()
+        action_ = np.argmax(qval_) #B
+        action = action_set[action_]
+        if display:
+            print('Move #: %s; Taking action: %s' % (i, action))
+        test_game.makeMove(action)
+        state_ = test_game.board.render_np().reshape(1,64) + np.random.rand(1,64)/10.0
+        state = torch.from_numpy(state_).float()
+        if display:
+            print(test_game.display())
+        reward = test_game.reward()
+        if reward != -1:
+            if reward > 0:
+                status = 2
+                if display:
+                    print("Game won! Reward: %s" % (reward,))
+            else:
+                status = 0
+                if display:
+                    print("Game LOST. Reward: %s" % (reward,))
+        i += 1
+        if (i > 15):
+            if display:
+                print("Game lost; too many moves.")
+            break
+    
+    win = True if status == 2 else False
+    return win
