+++ requirements.txt
@@ -26,7 +26,7 @@ pytz==2018.5
 requests==2.19.1
 simplegeneric==0.8.1
 six==1.11.0
-torch==0.4.0
+torch==1.9.0
 traitlets==4.3.2
 urllib3==1.23
 wcwidth==0.1.7
